+++ packages/dialog/src/mwc-dialog-base.ts
@@ -47,6 +47,8 @@ export class DialogBase extends BaseElement {
   // undefined" error in browsers that don't define it (e.g. Edge and IE11).
   @query('slot[name="secondaryAction"]') protected secondarySlot!: HTMLElement;
 
+  @query('#contentSlot') protected contentSlot!: HTMLElement;
+
   @query('.mdc-dialog__content') protected contentElement!: HTMLDivElement;
 
   @query('.mdc-container') protected conatinerElement!: HTMLDivElement;
@@ -118,7 +120,58 @@ export class DialogBase extends BaseElement {
   }
 
   protected getInitialFocusEl(): HTMLElement|null {
-    return this.querySelector(`[${this.initialFocusAttribute}]`);
+    const initFocusSelector = `[${this.initialFocusAttribute}]`;
+
+    // only search light DOM. This typically handles all the cases
+    const lightDomQs = this.querySelector(initFocusSelector);
+
+    if (lightDomQs) {
+      return lightDomQs as HTMLElement;
+    }
+
+    // if not in light dom, search each flattened distributed node.
+    const primarySlot = this.primarySlot as HTMLSlotElement;
+    const primaryNodes = primarySlot.assignedNodes({flatten: true});
+    const primaryFocusElement = this.searchNodeTreesForAttribute(
+        primaryNodes, this.initialFocusAttribute);
+    if (primaryFocusElement) {
+      return primaryFocusElement;
+    }
+
+    const secondarySlot = this.secondarySlot as HTMLSlotElement;
+    const secondaryNodes = secondarySlot.assignedNodes({flatten: true});
+    const secondaryFocusElement = this.searchNodeTreesForAttribute(
+        secondaryNodes, this.initialFocusAttribute);
+    if (secondaryFocusElement) {
+      return secondaryFocusElement;
+    }
+
+
+    const contentSlot = this.contentSlot as HTMLSlotElement;
+    const contentNodes = contentSlot.assignedNodes({flatten: true});
+    const initFocusElement = this.searchNodeTreesForAttribute(
+        contentNodes, this.initialFocusAttribute);
+    return initFocusElement;
+  }
+
+  private searchNodeTreesForAttribute(nodes: Node[], attribute: string):
+      HTMLElement|null {
+    for (const node of nodes) {
+      if (!(node instanceof HTMLElement)) {
+        continue;
+      }
+
+      if (node.hasAttribute(attribute)) {
+        return node;
+      } else {
+        const selection = node.querySelector(`[${attribute}]`);
+        if (selection) {
+          return selection as HTMLElement;
+        }
+      }
+    }
+
+    return null;
   }
 
   protected createAdapter(): MDCDialogAdapter {
@@ -201,7 +254,7 @@ export class DialogBase extends BaseElement {
         <div class="mdc-dialog__surface">
           ${heading}
           <div id="content" class="mdc-dialog__content">
-            <slot></slot>
+            <slot id="contentSlot"></slot>
           </div>
           <footer
               id="actions"